Move WindowSize from ui module
