+++ src/bin/client/graphics/ui/ui_pass.rs
@@ -1,12 +1,12 @@
 use bytemuck::{Pod, Zeroable};
 use crossbeam_channel::Sender;
+use unnamed_rts::resources::WindowSize;
 use wgpu::{
     include_spirv,
     util::{BufferInitDescriptor, DeviceExt},
     CommandBuffer, Device,
 };
 
-use super::ui_context::WindowSize;
 
 #[derive(Debug)]
 enum BufferType {
